+++ server/src/asm/syntaxes/64tass.ts
@@ -24,6 +24,7 @@ export class Tass64Syntax extends SyntaxDef {
   public allowLabelTrailingColon = true
   public allowIndentedAssignment = true
   public allowLineContinuation = false
+  public stringEscapeChars = ""
   public scopeSeparator = "."
   public defaultOrg = 0x0800        // TODO: choose correct value