+++ server/src/asm/syntaxes/lisa.ts
@@ -30,7 +30,7 @@ export class LisaSyntax extends SyntaxDef {
   public allowLabelTrailingColon = false
   public allowIndentedAssignment = false
   public allowLineContinuation = false
-  public allowStringEscapes = false
+  public stringEscapeChars = ""
   public scopeSeparator = ""
   public defaultOrg = 0x0800        // TODO: choose correct value